Moved Boards, added json, refactor
moved boards into their own folders

Added json for each board

WIP
