+++ src/examples/fill-mask/README.md
@@ -1,0 +1,5 @@
+Board which calls the Hugging Face Interface Fill Mask Endpoint.
+
+For more Details See: https://huggingface.co/docs/api-inference/detailed_parameters#fill-mask-task 
+
+To test on Breaboard Web : https://breadboard-ai.web.app/?mode=build&board=https%3A%2F%2Fraw.githubusercontent.com%2FExaDev%2Fbreadboard-examples%2FHugging-Face-Clean-History%2Fsrc%2Fexamples%2Ffill-mask%2Fboard.json
