+++ src/examples/question-answering/board.json
@@ -1,0 +1,200 @@
+{
+	"title": "Hugging Face Question Answering Board",
+	"description": "Board which calls the Hugging Face Question Answering Endpoint",
+	"edges": [
+		{
+			"from": "main",
+			"to": "output-2",
+			"out": "output",
+			"in": "output"
+		},
+		{
+			"from": "fetch-5",
+			"to": "main",
+			"out": "*",
+			"in": ""
+		},
+		{
+			"from": "fn-3",
+			"to": "fetch-5",
+			"out": "auth",
+			"in": "headers"
+		},
+		{
+			"from": "fn-4",
+			"to": "fetch-5",
+			"out": "payload",
+			"in": "body"
+		},
+		{
+			"from": "query",
+			"to": "fn-3",
+			"out": "apiKey",
+			"in": "key"
+		},
+		{
+			"from": "query",
+			"to": "fn-4",
+			"out": "*",
+			"in": ""
+		}
+	],
+	"nodes": [
+		{
+			"id": "output-2",
+			"type": "output",
+			"configuration": {
+				"schema": {
+					"type": "object",
+					"properties": {
+						"output": {
+							"type": "string",
+							"title": "output"
+						}
+					}
+				}
+			}
+		},
+		{
+			"id": "main",
+			"type": "output",
+			"configuration": {}
+		},
+		{
+			"id": "fetch-5",
+			"type": "fetch",
+			"configuration": {
+				"method": "POST",
+				"url": "https://api-inference.huggingface.co/models/deepset/roberta-base-squad2"
+			}
+		},
+		{
+			"id": "fn-3",
+			"type": "invoke",
+			"configuration": {
+				"$board": "#fn-3"
+			}
+		},
+		{
+			"id": "fn-4",
+			"type": "invoke",
+			"configuration": {
+				"$board": "#fn-4"
+			}
+		},
+		{
+			"id": "query",
+			"type": "input",
+			"configuration": {
+				"schema": {
+					"title": "Hugging Face Schema For Question Answering",
+					"properties": {
+						"question": {
+							"type": "string",
+							"title": "question",
+							"default": "What is my name?",
+							"description": "The data to send to the hugging face api question answering endpoint"
+						},
+						"apiKey": {
+							"type": "string",
+							"title": "apiKey",
+							"default": "myKey",
+							"description": "The hugging face api key"
+						},
+						"context": {
+							"type": "string",
+							"title": "context",
+							"default": "My name is Clara and I live in Berkeley.",
+							"description": "context for the question being asked"
+						},
+						"use_cache": {
+							"type": "boolean",
+							"title": "use_cache",
+							"default": "true",
+							"description": "Boolean. There is a cache layer on the inference API to speedup requests we have already seen. Most models can use those results as is as models are deterministic (meaning the results will be the same anyway). However if you use a non deterministic model, you can set this parameter to prevent the caching mechanism from being used resulting in a real new query"
+						},
+						"wait_for_model": {
+							"type": "boolean",
+							"title": "wait_for_model",
+							"default": "false",
+							"description": " Boolean. If the model is not ready, wait for it instead of receiving 503. It limits the number of requests required to get your inference done. It is advised to only set this flag to true after receiving a 503 error as it will limit hanging in your application to known places"
+						}
+					}
+				},
+				"type": "string"
+			}
+		}
+	],
+	"graphs": {
+		"fn-3": {
+			"edges": [
+				{
+					"from": "fn-3-input",
+					"to": "fn-3-run",
+					"out": "*"
+				},
+				{
+					"from": "fn-3-run",
+					"to": "fn-3-output",
+					"out": "*"
+				}
+			],
+			"nodes": [
+				{
+					"id": "fn-3-input",
+					"type": "input",
+					"configuration": {}
+				},
+				{
+					"id": "fn-3-run",
+					"type": "runJavascript",
+					"configuration": {
+						"code": "function fn_3(inputs) {const key=inputs.key;const auth={Authorization:`Bearer ${key}`};return{auth}}",
+						"name": "fn_3",
+						"raw": true
+					}
+				},
+				{
+					"id": "fn-3-output",
+					"type": "output",
+					"configuration": {}
+				}
+			]
+		},
+		"fn-4": {
+			"edges": [
+				{
+					"from": "fn-4-input",
+					"to": "fn-4-run",
+					"out": "*"
+				},
+				{
+					"from": "fn-4-run",
+					"to": "fn-4-output",
+					"out": "*"
+				}
+			],
+			"nodes": [
+				{
+					"id": "fn-4-input",
+					"type": "input",
+					"configuration": {}
+				},
+				{
+					"id": "fn-4-run",
+					"type": "runJavascript",
+					"configuration": {
+						"code": "function fn_4(input) {const{question,context}=input;const payload={\"inputs\":{question,context}};return{payload}}",
+						"name": "fn_4",
+						"raw": true
+					}
+				},
+				{
+					"id": "fn-4-output",
+					"type": "output",
+					"configuration": {}
+				}
+			]
+		}
+	}
+}
