+++ src/examples/text-classification/index.ts
@@ -1,19 +1,13 @@
 import { board, base, code } from "@google-labs/breadboard";
 import { core } from "@google-labs/core-kit";
-import { HuggingFaceTask } from "./types.js";
+import path from "path";
+import fs from "fs";
 
-const soruceSentenceSchema = {
+const inputsSchema = {
     type: "string",
     title: "inputs",
-    default: "That is a happy person",
-    description: "The data to send to the hugging face api sentence similarity endpoint"
-};
-
-const sentencesSchema = {
-    type: "string",
-    title: "inputs",
-    default: "[That is a happy dog, That is a very happy person,Today is a sunny day]",
-    description: "A list of sentences to compare the source sentence to"
+    default: "I like you. I love you",
+    description: "The data to send to the hugging face api text classification endpoint"
 };
 
 const keySchema = {
@@ -37,49 +31,33 @@ const waitForModelSchema = {
     description: " Boolean. If the model is not ready, wait for it instead of receiving 503. It limits the number of requests required to get your inference done. It is advised to only set this flag to true after receiving a 503 error as it will limit hanging in your application to known places"
 };
 
-export type HuggingFaceSentenceSimilarityRawParams = {
-    source_sentence: string
-    sentences: string[]
-
-    use_cache: boolean
-    wait_for_model: boolean
-};
-
-export type HuggingFaceSentenceSimilarityParams = {
-    inputs: {
-        source_sentence: string
-        sentences: string[]
-    }
+export type HuggingFaceTextClassificationRawParams = {
+    inputs: string
 };
 
 const authenticate = code<{ key: string }>((inputs) => {
     const key = inputs.key
-    const auth = { Authorization: `Bearer ${key}` }
+    const auth = { Authorization: `Bearer ${key}` };
 
     return { auth };
 });
 
-const handleParams = code<{ input: HuggingFaceSentenceSimilarityRawParams }>((input) => {
-    const { source_sentence, sentences } = input
+const handleParams = code<{ inputs: string }>((input) => {
+    const { inputs } = input
 
-    const payload: HuggingFaceSentenceSimilarityParams = {
-        "inputs": {
-            source_sentence: source_sentence,
-            sentences: sentences,
-        }
-    };
+    const request: HuggingFaceTextClassificationRawParams = { inputs: inputs };
+    const payload = JSON.stringify(request);
 
     return { payload }
-});
+})
 
-const huggingFaceBoardSentenceSimilarity = board(() => {
+const serialized = await board(() => {
     const inputs = base.input({
         $id: "query",
         schema: {
-            title: "Hugging Face Schema For Sentence Similarity",
+            title: "Hugging Face Schema For Text Classification",
             properties: {
-                soruce_sentence: soruceSentenceSchema,
-                sentences: sentencesSchema,
+                inputs: inputsSchema,
                 apiKey: keySchema,
                 use_cache: useCacheSchema,
                 wait_for_model: waitForModelSchema
@@ -88,10 +66,10 @@ const huggingFaceBoardSentenceSimilarity = board(() => {
         type: "string",
     });
 
-    const task = HuggingFaceTask.sentenceSimilarity
+    const task = "https://api-inference.huggingface.co/models/distilbert-base-uncased-finetuned-sst-2-english"
     const output = base.output({ $id: "main" });
 
-    const { auth } = authenticate({ key: inputs.apiKey as unknown as string })
+    const { auth } = authenticate({ key: inputs.apiKey as unknown as string });
     const { payload } = handleParams(inputs);
 
     const response = core.fetch({
@@ -103,12 +81,13 @@ const huggingFaceBoardSentenceSimilarity = board(() => {
 
     response.to(output);
     return { output }
+}).serialize({
+    title: "Hugging Face Text Classification",
+    description: "Board which calls the Hugging Face Text Classification Endpoint"
 });
 
-const source_sentence = "That is a happy person"
-const sentences = ["That is a happy dog", "That is a very happy person", "Today is a sunny day"]
-
-console.log(
-    JSON.stringify(await huggingFaceBoardSentenceSimilarity({ source_sentence: source_sentence, sentences: sentences, apiKey: "myAPiKey" }), null, 2)
+fs.writeFileSync(
+    path.join(".", "board.json"),
+    JSON.stringify(serialized, null, "\t")
 );
