+++ src/examples/translation-RU-ENG/index.ts
@@ -1,12 +1,14 @@
 import { board, base, code } from "@google-labs/breadboard";
 import { core } from "@google-labs/core-kit";
-import { HuggingFaceTask } from "./types.js";
+import path from "path";
+import fs from "fs";
+
 
 const inputsSchema = {
     type: "string",
-    title: "inputs",
-    default: "Hi, I recently bought a device from your company but it is not working as advertised and I would like to get reimbursed!",
-    description: "The data to send to the hugging face api labelling endpoint"
+    title: "data",
+    default: "Меня зовут Вольфганг и я живу в Берлине",
+    description: "The data to send to the Hugging Face api Translation API (Russian To English)"
 };
 
 const keySchema = {
@@ -14,20 +16,6 @@ const keySchema = {
     title: "apiKey",
     default: "myKey",
     description: "The hugging face api key"
-};
-
-const candidateLabelsSchema = {
-    type: "list",
-    title: "candidate_lebels",
-    default: "[refund, legal, faq]",
-    description: "The labels to mark the input"
-};
-
-const multiLabelSchema = {
-    type: "boolean",
-    title: "multi_label",
-    default: "true",
-    description: "Flag to indicate if multi labels are allowed"
 };
 
 const useCacheSchema = {
@@ -45,49 +33,39 @@ const waitForModelSchema = {
 };
 
 
-export type HuggingFaceLabellingRawParams = {
-    inputs: string;
-    candidate_labels: string[];
-    multi_label: boolean;
-    use_cache: boolean;
-    wait_for_model: boolean;
+export type HuggingFaceTranslationRawParams = {
+    inputs: string
+    use_cache: boolean
+    wait_for_model: boolean
 };
 
-export type HuggingFaceLabellingParams = {
-    inputs: string;
+export type HuggingFaceTranslationParams = {
+    inputs: string
     parameters: {
-        candidate_labels: string[];
-        multi_label: boolean;
-
         options: {
-            use_cache: boolean;
-            wait_for_model: boolean;
+            use_cache: boolean
+            wait_for_model: boolean
         }
     }
 };
 
-
 const authenticate = code<{ key: string }>((inputs) => {
     const key = inputs.key
-    const auth = { Authorization: `Bearer ${key}` }
+    const auth = { Authorization: `Bearer ${key}` };
 
     return { auth };
 });
 
-const handleParams = code<{ input: HuggingFaceLabellingRawParams }>((input) => {
+const handleParams = code<{ inputs: string, use_cache: boolean, wait_for_model:boolean }>((input) => {
     const {
         inputs,
-        candidate_labels,
-        multi_label,
         use_cache,
         wait_for_model
     } = input
 
-    const payload: HuggingFaceLabellingParams = {
+    const request: HuggingFaceTranslationParams = {
         inputs: inputs,
         parameters: {
-            candidate_labels: candidate_labels,
-            multi_label: multi_label,
             options: {
                 use_cache: use_cache,
                 wait_for_model: wait_for_model
@@ -95,19 +73,19 @@ const handleParams = code<{ input: HuggingFaceLabellingRawParams }>((input) => {
         }
     };
 
-    return { payload }
+    const payload = JSON.stringify(request);
+
+    return { payload };
 });
 
-const huggingFaceBoardTextLabelling = board(() => {
+const serialized = await board(() => {
     const inputs = base.input({
         $id: "query",
         schema: {
-            title: "Hugging Face Schema For text labelling",
+            title: "Hugging Face Schema For RUSSIAN TO ENGLISH TRANSLATION",
             properties: {
                 inputs: inputsSchema,
                 apiKey: keySchema,
-                candidate_labels: candidateLabelsSchema,
-                multi_label: multiLabelSchema,
                 use_cache: useCacheSchema,
                 wait_for_model: waitForModelSchema
             },
@@ -115,7 +93,7 @@ const huggingFaceBoardTextLabelling = board(() => {
         type: "string",
     });
 
-    const task = HuggingFaceTask.labelling
+    const task = "https://api-inference.huggingface.co/models/Helsinki-NLP/opus-mt-ru-en"
     const output = base.output({ $id: "main" });
 
     const { auth } = authenticate({ key: inputs.apiKey as unknown as string });
@@ -131,11 +109,12 @@ const huggingFaceBoardTextLabelling = board(() => {
     response.to(output);
 
     return { output }
+}).serialize({
+    title: "Hugging Face Translation Russian to English Board",
+    description: "Board which calls the Hugging Face Translation Endpoint"
 });
 
-const data = "Hi, I recently bought a device from your company but it is not working as advertised and I would like to get reimbursed!"
-const candidate_labels = ["refund", "legal", "faq"]
-
-console.log(
-    JSON.stringify(await huggingFaceBoardTextLabelling({ inputs: data, candidate_labels: candidate_labels, multi_label: false, apiKey: "myAPiKey", use_cache: true, wait_for_model: true }), null, 2)
+fs.writeFileSync(
+    path.join(".", "board.json"),
+    JSON.stringify(serialized, null, "\t")
 );
